new report.jsx and new fileUpload
