+++ carpion-client/src/components/page/FileUpload/FileUpload.css
@@ -1,0 +1,95 @@
+/* FileUpload.css */
+
+.file-upload-section {
+  display: flex;
+  align-items: center;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  justify-content: space-between;
+  gap: 10px;
+}
+
+.file-input-hidden {
+  display: none;
+}
+
+.file-upload-button {
+  background-color: #35bdf2;
+  color: white;
+  padding: 10px 20px;
+  cursor: pointer;
+  font-size: 1rem;
+  border: none;
+  border-radius: 6px;
+  transition: background-color 0.2s;
+}
+
+.file-upload-button:hover {
+  background-color: #007bff;
+}
+
+.file-description {
+  margin-top: 5px;
+  color: #777;
+  font-size: 14px;
+}
+
+.selected-file-info {
+  font-size: 14px;
+  color: #555;
+}
+
+.file-delete-button {
+  background-color: #dc3545;
+  color: white;
+  border: none;
+  border-radius: 6px;
+  padding: 10px 15px;
+  cursor: pointer;
+  font-size: 1rem;
+  transition: background-color 0.2s;
+}
+
+.file-delete-button:hover {
+  background-color: #c82333;
+}
+
+.file-selector-group {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.file-preview-section {
+  margin-top: 20px;
+}
+
+.file-preview-section h4 {
+  font-size: 1.2rem;
+  margin-bottom: 10px;
+}
+
+.preview-container {
+  display: flex;
+  gap: 10px;
+  overflow-x: auto;
+}
+
+.preview-item {
+  border: 1px solid #ccc;
+  padding: 5px;
+  border-radius: 4px;
+}
+
+.preview-item img {
+  display: block;
+  max-width: 100px;
+  max-height: 100px;
+}
+
+.preview-item span {
+  display: block;
+  font-size: 0.9rem;
+  color: #777;
+  text-align: center;
+}
