+++ carpion-client/src/components/page/reportpage/ReportForm.css
@@ -1,0 +1,42 @@
+/* ReportForm.css */
+
+.write-page-container {
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 40px 20px;
+}
+
+.write-page-title {
+  font-size: 2rem;
+  font-weight: bold;
+  margin-bottom: 20px;
+}
+
+.post-title-input {
+  width: 100%;
+  padding: 12px;
+  font-size: 1.1rem;
+  border: 1px solid #ccc;
+  border-radius: 6px;
+  margin-bottom: 20px;
+}
+
+.submit-button {
+  padding: 10px 20px;
+  background-color: #007bff;
+  color: white;
+  font-size: 1rem;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  margin-left: 10px; 
+}
+
+.submit-button:hover {
+  background-color: #0056b3;
+}
+
+.toastui-editor-mode-switch {
+  display: none !important;
+}
