--- carpion-client/src/components/page/reportpage/ReportPage.css
@@ -1,111 +1,0 @@
-/* ReportPage.css */
-
-.write-page-container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 40px 20px;
-}
-
-.write-page-title {
-  font-size: 2rem;
-  font-weight: bold;
-  margin-bottom: 20px;
-}
-
-.post-title-input {
-  width: 100%;
-  padding: 12px;
-  font-size: 1.1rem;
-  border: 1px solid #ccc;
-  border-radius: 6px;
-  margin-bottom: 20px;
-}
-
-.submit-button {
-  padding: 10px 20px;
-  background-color: #007bff;
-  color: white;
-  font-size: 1rem;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-  margin-left: 10px; 
-}
-
-.submit-button:hover {
-  background-color: #0056b3;
-}
-
-.toastui-editor-mode-switch {
-  display: none !important;
-}
-
-.file-upload-section {
-  display: flex;
-  align-items: center;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  justify-content: space-between; 
-}
-
-.file-input-hidden {
-  display: none;
-}
-
-.file-upload-button {
-  background-color: #35bdf2;
-  color: white;
-  padding: 10px 20px;
-  cursor: pointer;
-  font-size: 1rem;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.file-upload-button:hover {
-  background-color: #007bff;
-}
-
-.file-description {
-  margin-top: 5px;
-  color: #777;
-  font-size: 14px;
-}
-
-.selected-file-info {
-  font-size: 14px;
-  color: #555;
-}
-
-.file-delete-button {
-  background-color: #dc3545; 
-  color: white;
-  border: none;
-  border-radius: 6px;
-  padding: 10px 15px;
-  cursor: pointer;
-  font-size: 1rem;
-  transition: background-color 0.2s; 
-}
-
-.file-delete-button:hover {
-  background-color: #c82333;
-}
-
-.file-upload-section {
-  display: flex;
-  align-items: center;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  justify-content: space-between; 
-  gap: 10px; 
-}
-
-.file-selector-group {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
